Provide bit-casted slot values from DebugTracePlayer.
Previously, the VariableData returned from a DebugTracePlayer contained
values in an int32_t, regardless of the slot's NumberKind. We had a
helper function which could stringize the bits, but otherwise the caller
was responsible for bit-casting the value manually.

Now, the DebugTracePlayer will automatically manage bit-casting for the
caller. The value returned in the VariableData is now a double (so it
is able to store an int32, uint32, or float at full precision).

This change was inspired by the recent Typescript port. (The value in
the Typescript VariableData uses a compound `number | boolean` type, so
it is able to fully represent any slot value natively.)

Bug: skia:12666
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/489896
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>

+++ src/sksl/tracing/SkVMDebugTracePlayer.h
@@ -74,9 +74,9 @@ public:
 
     /** Returns variables from a stack frame, or from global scope. */
     struct VariableData {
-        int      fSlotIndex;
-        bool     fDirty;  // has this slot been written-to since the last step call?
-        int32_t  fValue;  // caller must type-pun bits to float/bool based on slot type
+        int     fSlotIndex;
+        bool    fDirty;  // has this slot been written-to since the last step call?
+        double  fValue;  // value in slot (with type-conversion applied)
     };
     std::vector<VariableData> getLocalVariables(int stackFrameIndex) const;
     std::vector<VariableData> getGlobalVariables() const;
